requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
llama-index>=0.10.0
llama-parse>=0.4.0
llama-index-llms-gemini>=0.1.6
llama-index-embeddings-openai>=0.1.0
llama-index-vector-stores-chroma>=0.1.0
chromadb>=0.4.0
openai>=1.10.0
google-genai>=0.2.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
aiosqlite>=0.19.0
streamlit>=1.30.0
requests>=2.31.0
